requirements.txt
fastapi
uvicorn[standard]
openai
requests
python-dotenv
pandas
sqlalchemy
psycopg2-binary
faiss-cpu
sentence-transformers
langchain
streamlit
matplotlib
plotly
python-multipart
pydantic
python-dateutil
google-genai